requirements.txt
python-dotenv==1.0.0
atlassian-python-api==3.41.6
beautifulsoup4==4.12.3
bs4==0.0.2
BeautifulSoup4==4.12.3
